scenarios/S05-service-validation/solution/app/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pyodbc==4.0.39
azure-identity==1.17.1
python-dotenv==1.0.1
requests==2.32.3
dnspython==2.6.1
mpmath==1.3.0
